requirements.txt
mysqlclient==1.3.13
django-ckeditor==5.6.1